+++ crates/azure_devops_types/src/lib.rs
@@ -4,6 +4,7 @@ mod azure_devops_entra_user_descriptor;
 mod azure_devops_group;
 mod azure_devops_group_member;
 mod azure_devops_license_entitlement;
+mod azure_devops_agent_package;
 mod azure_devops_organization_name;
 mod azure_devops_organization_url;
 mod azure_devops_project;
@@ -38,6 +39,7 @@ pub mod prelude {
     pub use crate::azure_devops_group::*;
     pub use crate::azure_devops_group_member::*;
     pub use crate::azure_devops_license_entitlement::*;
+    pub use crate::azure_devops_agent_package::*;
     pub use crate::azure_devops_organization_name::*;
     pub use crate::azure_devops_organization_url::*;
     pub use crate::azure_devops_project::*;