+++ crates/entrypoint/src/cli/command/azure_devops/agent_package/list/mod.rs
@@ -1,0 +1,2 @@
+mod azure_devops_agent_package_list_cli;
+pub use azure_devops_agent_package_list_cli::*;