+++ crates/entrypoint/src/cli/command/azure_devops/agent_package/mod.rs
@@ -1,0 +1,4 @@
+mod azure_devops_agent_package_cli;
+pub mod list;
+pub mod show_newest;
+pub use azure_devops_agent_package_cli::*;