+++ crates/entrypoint/src/cli/command/azure_devops/agent_package/show_newest/mod.rs
@@ -1,0 +1,2 @@
+mod azure_devops_agent_package_show_newest_cli;
+pub use azure_devops_agent_package_show_newest_cli::*;